swagger and new entity added

+++ src/main/java/com/ahmetyasin/demogetir/repository/BaseRepository.java
@@ -1,13 +1,12 @@
 package com.ahmetyasin.demogetir.repository;
 
-import com.ahmetyasin.demogetir.entity.Entity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.NoRepositoryBean;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 @NoRepositoryBean
-public interface BaseRepository<T> extends PagingAndSortingRepository<T, Integer>, JpaSpecificationExecutor<T>, JpaRepository<T,Integer> {
+public interface BaseRepository<T> extends PagingAndSortingRepository<T, Long>, JpaSpecificationExecutor<T>, JpaRepository<T,Long> {
 
     @Override
     <S extends T> S save(S entity);
